Times in SFML are now Uint32 milliseconds instead of float seconds.
(SFML's corresponding commit: #e4c6c30e0be80eacb7e6)

+++ decl.pxd
@@ -153,7 +153,7 @@ cdef extern from "SFML/Graphics.hpp" namespace "sf":
 
     cdef cppclass Clock:
         Clock()
-        float GetElapsedTime()
+        Uint32 GetElapsedTime()
         void Reset()
 
     cdef cppclass Color:
@@ -395,7 +395,7 @@ cdef extern from "SFML/Graphics.hpp" namespace "sf":
         void Draw(Drawable&, Shader&)
         void EnableKeyRepeat(bint)
         View& GetDefaultView()
-        float GetFrameTime()
+        Uint32 GetFrameTime()
         Input& GetInput()
         unsigned int GetHeight()
         ContextSettings& GetSettings()
